page.tsx moved back to /Data
