+++ app/Data/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React from "react";
 import Link from "next/link";
-import { contacts } from "../Data/contacts";
+import { contacts } from "./contacts";
 import AllContactsList from "../../components/AllContactsList";
 import SearchBar from "../../components/SearchBar";
 import AddContactsButton from "../../components/AddContactsButton";
